fix(EoMT): ensure model correctness for training and evaluation
Fixes three critical EoMT integration issues:

1. generate() now carries eomt_images/eomt_meta through prepare_inputs_labels_for_multimodal,
   ensuring eval uses same EoMT path as training (fixes train/eval mismatch).

2. Learnable OBJ_INFO and object-type embeddings now eagerly registered before optimizer
   creation, guaranteeing they enter param groups (fixes silent non-trainable params).

3. EoMT pooling fallback reasons now explicit and aggregated, especially for anyres/multi-patch
   cases, with rank0 warnings to surface when object-block append is skipped.

+++ llava/model/multimodal_eomt/mask_pooler.py
@@ -293,7 +293,7 @@ class MaskGuidedPooler(nn.Module):
         weights = weights / weights.sum(dim=(-1, -2), keepdim=True).clamp_min(self.eps)
 
         grid_tokens_flat = grid_tokens.view(batch_size, grid_h * grid_w, feat_dim)
-        weights_flat = weights.view(batch_size, k, grid_h * grid_w)
+        weights_flat = weights.view(batch_size, k, grid_h * grid_w).to(grid_tokens_flat.dtype)
         pooled_tokens = torch.einsum("bkh,bhd->bkd", weights_flat, grid_tokens_flat)
 
         return {
